requirements.txt
certifi==2017.4.17
chardet==3.0.3
click==6.7
decorator==4.0.11
Django==1.11.2
dnspython==1.15.0
future==0.16.0
geocoder==1.22.6
httplib2==0.10.3
idna==2.5
MySQL-python==1.2.5
olefile==0.44
pyasn1==0.2.3
pyasn1-modules==0.0.9
python-whois==0.6.5
pytz==2017.2
ratelim==0.1.6
requests==2.17.3
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.0
